field: move subset-sum helpers from binius-math to binius-field
Move expand_subset_sums_array and expand_subset_sums (and their tests)
from the binius_math::span module into binius_field::util, and delete
the now-empty span module.

These helpers expand a set of field elements into all of their subset
sums, the lookup tables behind the Method of Four Russians. They operate
purely on Field elements with no dependency on anything in binius-math,
and binius-field is the lower-level crate, so util is their natural home.

The move also lets binius-field reuse expand_subset_sums_array directly:
the Itoh-Tsujii inverter had its own private copy of the function, which
is now removed in favor of the shared util one.

Update sites:
- binius_math::span deleted; pub mod span removed from the crate root.
- prover/ring_switch.rs imports expand_subset_sums_array from
  binius_field::util instead of binius_math::span.
- arch/portable/arithmetic/itoh_tsujii.rs drops its duplicated definition
  and imports crate::util::expand_subset_sums_array.
