+++ crates/math/src/lib.rs
@@ -23,7 +23,6 @@ pub mod matrix;
 pub mod multilinear;
 pub mod ntt;
 pub mod reed_solomon;
-pub mod span;
 pub mod tensor_algebra;
 #[cfg(feature = "test-utils")]
 pub mod test_utils;
